requirements.txt
streamlit
torch
torchvision
torchaudio
transformers
pillow
gTTS
easyocr
accelerate